+++ app/Http/Controllers/SiteController.php
@@ -35,6 +35,24 @@ class SiteController extends Controller
     }
 
     /**
+     * Display a list of branches on the given site's repository.
+     *
+     * @param Request $request
+     * @param Site    $site
+     *
+     * @return Response
+     */
+    public function branches(Request $request, Site $site)
+    {
+        $cmd = "git ls-remote --heads '%s' | sed 's^.*refs/heads/^^'";
+        $repo = $request->query('repo', $site->source_repo);
+
+        exec(sprintf($cmd, $repo), $branches);
+
+        return $branches;
+    }
+
+    /**
      * Pull the latest commit from Git.
      *
      * @param \Servidor\Site $site